refactor: product 패키징 변경
