+++ service/store/src/main/java/com/webest/store/product/domain/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package com.webest.store.product.domain;
+package com.webest.store.product.domain.repository;
 
+import com.webest.store.product.domain.model.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
